requirements.txt
playwright
langchain-google-genai
python-dotenv
requests
pillow
openai
google-generativeai
